requirements.txt
pandas
requests
yake
python-dotenv
openpyxl
